+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+- Upload attachments (currently image-only)
 
 ## [0.1.4] - 2022-07-16
 ### Added
@@ -16,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Issue where messages sent from other device will land in the "Note to self" disregarding the recipient.
 - Contacts with unknown name were displayed as "Note to self" in the channel list. Now their phone numbers are displayed.
-- Maybe fix a crash where the application did not respond any more.
+- Maybe fix a crash where the application did not respond any more. (Update: The patch did not fix this issue. Still investigating)
 - Maybe fix a crash when the sled database was locked causing the window not to open.
 
 ### Development